add dbContext

+++ HR_Management.Application/Features/LeaveAllocations/Handlers/Commands/DeleteLeaveAllocationCommandHandler.cs
@@ -1,5 +1,8 @@
-﻿using HR_Management.Application.Features.LeaveAllocations.Requests.Commands;
+﻿using FluentValidation.Results;
+using HR_Management.Application.Execptions;
+using HR_Management.Application.Features.LeaveAllocations.Requests.Commands;
 using HR_Management.Application.Persistance.Contracts;
+using HR_Management.Domain;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -19,6 +22,8 @@ namespace HR_Management.Application.Features.LeaveAllocations.Handlers.Commands
         public async Task Handle(DeleteLeaveAllocationCommand request, CancellationToken cancellationToken)
         {
             var leaveAllocation =await _leaveAllocationRepository.Get(request.Id);
+            if(leaveAllocation == null)
+                throw new NotFoundExecption(nameof(LeaveAllocation),request.Id);
             await _leaveAllocationRepository.Delete(leaveAllocation);
         }
     }
