+++ HR_Management.Application/Features/LeaveRequests/Handlers/Commands/DeleteLeaveRequestCommandHandler.cs
@@ -1,5 +1,7 @@
-﻿using HR_Management.Application.Features.LeaveRequests.Requests.Commands;
+﻿using HR_Management.Application.Execptions;
+using HR_Management.Application.Features.LeaveRequests.Requests.Commands;
 using HR_Management.Application.Persistance.Contracts;
+using HR_Management.Domain;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -20,6 +22,8 @@ namespace HR_Management.Application.Features.LeaveRequests.Handlers.Commands
         public async Task Handle(DeleteLeaveRequestCommand request, CancellationToken cancellationToken)
         {
             var leaveRequest =await _leaveRequestRepository.Get(request.Id);
+            if (leaveRequest == null)
+                throw new NotFoundExecption(nameof(LeaveRequest), request.Id);
             await _leaveRequestRepository.Delete(leaveRequest);
         }
     }
